+++ src/main/java/ir/tic/clouddc/event/EventLandingForm.java
@@ -4,18 +4,17 @@ import ir.tic.clouddc.center.Center;
 import ir.tic.clouddc.center.Location;
 import ir.tic.clouddc.resource.Device;
 import lombok.Data;
-import lombok.NoArgsConstructor;
 import org.springframework.web.multipart.MultipartFile;
 
 @Data
 public class EventLandingForm {
-    private int eventCategoryId;
+    private Integer eventCategoryId;
     private String target;
     private Long locationId;
     private Location location;
-    private short centerId;
+    private Integer centerId;
     private Center center;
-    private short utilizerId;
+    private Integer utilizerId;
     private String serialNumber;
     private Device device;
     private String description;
